pyproject.toml
[build-system]
requires = ["setuptools>=66.0.0", "wheel"]
build-backend = "setuptools.build_meta:__legacy__"

[project]
name = "pygnssutils"
authors = [{ name = "semuadmin", email = "semuadmin@semuconsulting.com" }]
maintainers = [{ name = "semuadmin", email = "semuadmin@semuconsulting.com" }]
description = "GNSS Command Line Utilities"
version = "1.1.1"
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Operating System :: OS Independent",
    "Development Status :: 5 - Production/Stable",
    "Environment :: MacOS X",
    "Environment :: Win32 (MS Windows)",
    "Environment :: X11 Applications",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: End Users/Desktop",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: BSD License",
    "Topic :: Utilities",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: GIS",
]

dependencies = [
    "certifi>=2024.0.0",
    "paho-mqtt>=1.6.0",  # waiting for >=2.0.0 in Conda
    "pyserial>=3.5",
    "pyspartn>=1.0.2",
    "pyubx2>=1.2.45",
]

[project.scripts]
gnssstreamer = "pygnssutils.gnssstreamer_cli:main"
gnssdump = "pygnssutils.gnssstreamer_cli:main"           # for backwards compatibility
gnssserver = "pygnssutils.gnssserver_cli:main"
gnssntripclient = "pygnssutils.gnssntripclient_cli:main"
gnssmqttclient = "pygnssutils.gnssmqttclient_cli:main"
ubxsetrate = "pygnssutils.ubxsetrate:main"
ubxsave = "pygnssutils.ubxsave:main"
ubxload = "pygnssutils.ubxload:main"
ubxsimulator = "pygnssutils.ubxsimulator_cli:main"
ubxcompare = "pygnssutils.ubxcompare:main"

[project.urls]
homepage = "https://github.com/semuconsulting/pygnssutils"
documentation = "https://www.semuconsulting.com/pygnssutils/"
repository = "https://github.com/semuconsulting/pygnssutils"
changelog = "https://github.com/semuconsulting/pygnssutils/blob/master/RELEASE_NOTES.md"

[project.optional-dependencies]
deploy = ["build", "pip", "setuptools >= 66.0", "wheel"]
test = [
    "bandit",
    "black",
    "isort",
    "pylint",
    "pytest",
    "pytest-cov",
    "Sphinx",
    "sphinx-rtd-theme",
]

[tool.black]
target-version = ['py38']

[tool.isort]
py_version = 38
profile = "black"

[tool.bandit]
exclude_dirs = ["docs", "examples", "tests"]
skips = [
    "B104",
    "B105",
    "B307",
    "B311",
] # bind 0.0.0.0; hard coded pwd, use of eval, random gen

[tool.pylint]
jobs = 0
reports = "y"
recursive = "y"
py-version = "3.8"
fail-under = "9.8"
fail-on = "E,F"
clear-cache-post-run = "y"
disable = """
    bad-inline-option,
    deprecated-pragma,
    duplicate-code,
    file-ignored,
    locally-disabled,
    logging-fstring-interpolation,
    raw-checker-failed,
    suppressed-message,
    too-few-public-methods,
    too-many-instance-attributes,
    use-symbolic-message-instead,
    useless-suppression
"""

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "--cov --cov-report html --cov-fail-under 30"
pythonpath = ["src"]

[tool.coverage.run]
source = ["src"]
